requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
groq==0.4.2
pypdf==4.0.1
python-dotenv==1.0.1
Pillow==10.2.0
Werkzeug==3.0.1
gunicorn==21.2.0
